wip: Adding archive feature
wip: Fix conditional for archive prefix

If the prefix doesn't have a / at the end, add one.
Unless the prefix is empty (which is allowed), then don't do anything.

wip: pop up the extension menu every time

Validating the extension is easy in the positive case, but too difficult
in the negative case.